Feeds plugin updating
