--- __replace__/FeedsEmptyPanel.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import FeedsPanel from '../../../theme/plugins/Feeds/FeedsProfilePanel';
-
-const FeedsEmptyPanel = props => {
-    return (
-      <FeedsPanel {...props} />
-    );
-};
-
-export default FeedsEmptyPanel;
