--- __replace__/FeedsSelectors.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import FeedsCheckboxPanel from '../../../theme/plugins/Feeds/FeedsCheckboxPanel';
-
-const FeedsSelectors = props => {
-  return (
-    <FeedsCheckboxPanel {...props} />
-  );
-};
-
-export default FeedsSelectors;
