--- __replace__/non-core-utils.js
@@ -1,4 +1,0 @@
-import React from 'react';
-import { fetchFeedsOnMount } from '../../theme/plugins/Feeds/feeds-utils';
-
-export const pluginFeedsOnMount = fetchFeedsOnMount;
